Copy music from server
(CIFS only) if 'musicsharename' config variable is set, music will
be automatically copied from that location when connected to the
archive. The music share is assumed to be on the save server as
the camera archive and use the same credentials, but the share or
folder must be different.
The music disk size needs to be configured to be large enough to
hold all the music that will be copied over. If new files are
placed in the share, they will be automatically copied over next
time the Pi connects.